Move thinking process into one thread
There are over 1500 thinking iterations for one way iteration.
Implementation of thinking iterations in one thread with way iterations there is no way to calculate algorithm time good.
Because most transporters are stay when only one is going.

So it's better to wait each other transporter when one of them is thinking.